修改类型错误

+++ src/Migration/Columns/Thinkphp/FloatType.php
@@ -6,6 +6,6 @@ class FloatType extends AbstractType
 {
     public function migrateColumn():string
     {
-        return $this->getParsedField('double', $this->columnOptions(0, $this->column->getPrecision(), $this->column->getScale()));
+        return $this->getParsedField('float', $this->columnOptions(0, $this->column->getPrecision(), $this->column->getScale()));
     }
 }
